put cards on the tray when a kit is added
Add feat event handling for collision check when adding kit or
component.
For that reason feat event firing is opened to play_session.js.
This decision is made based on the basic architectual policy
which is 'do complex things on the browser where user interacts
and propagate the resulting status.'

Adding component is an user interaction but it's outside of feats.
So the events related to adding / removing components need to
be fired from outside of feats.